Added basic display class to support creating a pygame window, as well as support painting to the screen during each frame.
Added a lot of new feature support for logging, including:
- RotatingFileHandler to limit file size to a configurable size (in kilobytes).
- Configurable Regular Expression pattern filters to allow fine tuning of the log messages to show.
- Add option to disable loggers in configuration file.
- Add organization to event logging, categorizing by the name of the event the log message pertains too.

Phase out support of environment.py, as the project root is added to pythonpath by default in pycharm.

Add code to Engine class to keep track of a shared instance of the Engine class, and make all persistent variables class-level, so they are shared across instances.